Refactor EJBDirectory so it can be bundled via Archive.addPackage(...)
