+++ testsuite/integration/clust/src/test/java/org/jboss/as/test/clustering/ejb/AbstractEJBDirectory.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.clustering;
+package org.jboss.as.test.clustering.ejb;
 
 import java.util.Properties;
